requirements/base.txt
beautifulsoup4==4.10.0
crispy-bootstrap5==0.6
dateparser==1.1.1
Django==4.0.4
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-datepicker-widget==0.1
django-bootstrap4==21.1
django-crequest==2018.5.11
django-crispy-forms==1.14.0
django-environ==0.8.1
django-extensions==3.1.5
django-htmx==1.9.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
drf-extensions==0.7.1
mysqlclient==2.1.0
numpy==1.22.3
pandas==1.4.1
pip-review==1.1.1
Pillow>=9.0.1
pytz==2021.3
requests==2.27.1
slippers==0.3.1
